move registerParachain to helpers
