+++ tests/0011-update-genesis-config.toml
@@ -19,7 +19,7 @@ chain = "rococo-local"
 
 [[parachains]]
 id = 100
-add_to_genesis = true
+add_to_genesis = false
 
   [parachains.collator]
   name = "collator01"
